--- src/components/common/IconText/index.ts
@@ -1,2 +1,0 @@
-export { IconText } from "./IconText";
-export type { IconTextProps } from "./IconText";
